+++ dev/server/backend/api/routers/__init__.py
@@ -1,0 +1,1 @@
+# Routers package
