+++ dev/server/backend/api/routers/debug/__init__.py
@@ -1,0 +1,1 @@
+# Debug routers package
