--- dev/server/backend/api/routers/func/ai-review.py
@@ -1,15 +1,0 @@
-from fastapi import APIRouter
-
-# Routerを作成
-router = APIRouter()
-
-@router.get("/func/ai-review")
-async def ai_review() -> dict:
-    """
-    AIレビューのエンドポイントにアクセスした際の処理を行います。
-
-    Returns:
-      dict: レスポンスとして返されるメッセージを含む辞書。
-    """
-    return {"message": "AIレビューのエンドポイントにアクセスしました"}
-
