+++ dev/server/backend/api/routers/func/ai_review.py
@@ -1,0 +1,19 @@
+from fastapi import APIRouter
+
+# Routerを作成
+router = APIRouter()
+
+@router.post("/func/ai-review")
+async def ai_review(
+    review_data: dict
+) -> dict:
+    """
+    AIレビューエンドポイント - レビュー用のデータを受け取り、AIによるレビュー結果を返します
+    """
+    # ここではAIレビューのロジックを実装する必要があります
+    # 今回はサンプルとして、受け取ったデータをそのまま返す処理を行います
+
+    # デバッグ用に受け取ったデータをログに出力
+    print(f"Received review data: {review_data}")
+
+    return {"review_result": "Sample review result", "original_data": review_data}
